--- code/portfolio/src/containers/Section/Section.scss
@@ -1,52 +1,0 @@
-@import '../../MediaQueries.scss';
-
-.Section {
-    width: 100%;
-    padding: 0 20px;
-    display: flex;
-
-    @include lg {
-        height: 965px;
-    }
-
-    .ParticleSection {
-        width: 100%;
-        position: absolute;
-        top: 20%;
-        left: 50%;
-        transform: translate(-50%, -50%);
-        @include lg {
-            top:50%;
-        }
-    }
-}
-
-.row-lg{
-    @include lg{
-        display: flex;
-        flex-direction: row;
-    }
-}
-
-.column-lg{
-    @include lg{
-        display: flex;
-        flex-direction: column;
-    }
-}
-
-.row-sm{
-    display: flex;
-    flex-direction: row;
-}
-
-.column-sm{
-    display: flex;
-    flex-direction: column;
-}
-
-.no-padding {
-    padding: 0;
-}
-
-
