+++ code/portfolio/src/containers/Sections/Sections.scss
@@ -1,0 +1,53 @@
+@import '../../MediaQueries.scss';
+
+.ParticleSection {
+
+    .Particles {
+        width:100%;
+        position: absolute;
+        top: 0%;
+        left:0%;
+    }
+    background-color: #05386b;
+    height:260px;
+    @include lg {
+        height:955px;
+    }
+}
+
+.ValuesSection {
+    width:100%;
+    height: 200px;
+    text-align: center;
+    @include lg {
+        height: 400px;
+    }
+}
+
+.row-lg {
+    @include lg {
+        display: flex;
+        flex-direction: row;
+    }
+}
+
+.column-lg {
+    @include lg {
+        display: flex;
+        flex-direction: column;
+    }
+}
+
+.row-sm {
+    display: flex;
+    flex-direction: row;
+}
+
+.column-sm {
+    display: flex;
+    flex-direction: column;
+}
+
+.no-padding {
+    padding: 0;
+}
